Added program for the Web Server
Receives HTTP request through a socket, parses to make sure it is correctly formatted, then attempts to send the file requested. Has 404 and 501 HTTP errors implemented to indicate when a file request is not available or when the HTTP request header is improperly formatted. Note that this code cannot be run unless connected to the AFS servers at NJIT.